add fix

+++ MiniHelpDesk/App/Services/RoleService.cs
@@ -4,6 +4,7 @@ using App.Repositories.interfaces;
 using App.Services.interfaces;
 using System;
 using System.Collections.Generic;
+using System.Data;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -19,21 +20,28 @@ namespace App.Services
             _roleRepository = roleRepository;
         }
 
-        public Task AddRole(string name)
+        public async Task AddRole(string name)
         {
             if (string.IsNullOrWhiteSpace(name))
             {
                 throw new ArgumentException("Role name cannot be null or empty.");
             }
 
-            Role role = 
+            var role = await _roleRepository.GetRoleByName(name);
 
+            if (role is not null)
+            {
+                throw new InvalidOperationException($"Role with name '{name}' already exists.");
+            }
+
+            role = new Role
+            {
+                Name = name
+            };
+
+            await _roleRepository.AddAsync(role);
         }
 
-        public Task<Role?> GetRoleById(int id)
-        {
-            throw new NotImplementedException();
-        }
 
         public async Task<Role?> GetRoleByName(string name)
         {
@@ -57,7 +65,38 @@ namespace App.Services
 
         public async Task RemoveRole(int id)
         {
+            var role = await _roleRepository.GetByIdAsync(id) ?? throw new InvalidOperationException($"Not found role with id: {id}");
+
+            if (role.Name == "Null"      ||
+                role.Name == "Admin"     ||
+                role.Name == "Techicial" ||
+                role.Name == "Requester")
+            {
+                throw new InvalidOperationException("Cannot remove the 'Null', 'Admin', 'Techicial' or 'Requester' role.");
+            }
             await _roleRepository.RemoveAndSetDeffaut(id);
+        }
+
+        public async Task UpdateRole(int id, string newName)
+        {
+            if (string.IsNullOrWhiteSpace(newName))
+            {
+                throw new ArgumentException("Role name cannot be null or empty.");
+            }
+
+            var role = await _roleRepository.GetByIdAsync(id) ?? throw new InvalidOperationException($"Not found roles"); ;
+
+            if (role.Name == "Null"      ||
+                role.Name == "Admin"     ||
+                role.Name == "Techicial" ||
+                role.Name == "Requester")
+            {
+                throw new InvalidOperationException("Cannot update the 'Null', 'Admin', 'Techicial' or 'Requester' role.");
+            }
+
+            role.Name = newName;
+
+            await _roleRepository.UpdateAsync(role);
         }
     }
 }
